Nlp cleanup (#33)
* Testing json stored vectors

testing vectors stored in json

* removed file

Remove file to save space

* Create write_to_json.py

Submit a question and store it

* Update parser.py

better model usage via full question lists instead of one at a time

* Update __init__.py

* Update basiccli.py

Changes the way it presents. Improvements to the overall function of CLI

* Update universalencoder.py

Added preprocess step

* Create t5.py

Text summarisation

* Create summarisation.py

Abstract method for various summarisation methods

* Update questionmatcher.py

boolean to check if the suggestions should be text or subject

* Update __init__.py

* Update __init__.py

* quick syntax fix

* Bug fixes

Added so the target model is passed in which allows the program to read the correct file and write to it

* Bug fixes

Models now only embed questions and not all the posts

* Update basiccli.py

added target model to pass to write_to_json

* Update __init__.py

Added checks so that a directory is created if no directory exists to store. Also checks to see if a file exists already.

* Update t5.py

bug fixes with how to pass in text

* bug fixes

bug fixes with passing text into model

* Update __init__.py